Moving an area by clicking on the map Display markers if they are in the area

+++ ShowMap/models.py
@@ -16,8 +16,8 @@ class Monument(models.Model):
     image = models.ImageField(upload_to="images/monuments_image")
     wiki = models.CharField(max_length=100)
     slug = models.PositiveIntegerField(unique=True)
-    longtitude = models.IntegerField(blank=False)
-    latitude = models.IntegerField(blank=False)
+    longtitude = models.FloatField(blank=False)
+    latitude = models.FloatField(blank=False)
     icon = models.FileField(upload_to="images/monuments_icon")
     category = models.ForeignKey(Category, on_delete=models.CASCADE, blank=True, related_name="category")
